factor out functions
