+++ src/utils/data-management/loadRoles.js
@@ -1,0 +1,5 @@
+import fetchAPI from 'utils/http/fetchAPI';
+
+const loadRoles = async (experimentId) => fetchAPI(`/v1/access/${experimentId}`);
+
+export default loadRoles;
